ANDROID: usb: gadget: adb: Add ADB function
Android Debug Bridge (adb) is a command line tool that lets
users communicate with a Android-powered device. It is used
mainly to debug applications and tranfer files. f_adb implements
the transport layer between the ADB Server (on the host) and the
ADBD daemon (on the device).

Signed-off-by: Mike Lockwood <lockwood@android.com>
Git-commit: 796cdfa202cda0c18da419c02cb10a3062d5009b
Git-repo: https://android.googlesource.com/kernel/common.git
Signed-off-by: Andy Gross <agross@codeaurora.org>