USB: dwc: Add a warning message about Vendor Device LMP event
The current code is not handling vendor device test LMP event but enables
it in the DEVTEN register.  Fix it by not enabling this event in DEVTEN
register.

However controllers prior to 2.30a revision has a bug, due to which a
vendor device test LMP event can not be filtered.  But this event is not
handled in the current code.  This is a special event and 8 bytes of data
will follow the event.  Handling this event is tricky when event buffer
is almost full. Moreover this event will not occur in normal scenario and
can only happen with special hosts in testing scenarios.  Add a warning
message to indicate that this event is received which means that event
buffer might have corrupted.

CRs-fixed: 371306
Change-Id: I2858a3f1d90966f168e7e999f613396db7b1a972
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>