hwmon: pm8xxx-adc: Remove wakelocks
The PM8xxx ADC interrupt is configured as a
wakeable interrupt. Therefore the system can suspend
and still have the end of conversion interrupt be
serviced.

Add a timeout in the wait_for_completion in case
the end of conversion event is not received. This
is to rule out any side affects if a missed interrupt
occurs.

Signed-off-by: Siddartha Mohanadoss <smohanad@codeaurora.org>
(cherry picked from commit 96ab781d36a44372a23a517cfb7551be104ef275)

Change-Id: I78fbe130b3e80b9c39ced7d8bf0c610b3b68b543
Signed-off-by: Sridhar Gujje <sgujje@codeaurora.org>