power: pm8921-charger: fix EOC worker not starting
There are situations when we observe that charging is in progress
but the end of charge (EOC) worker is not running. This leads to
problems like
-never detecting end of charge
-not adjusting vddmax
-not preventing suspend while charging.

The reason is that the bootloader could have started charging already,
causing a fastchg interrupt to not happen. If the fastchg interrupt is
missed the driver will never start EOC worker.

Check the real time status of the fastchg interrupt and if it is fast
charging invoke the handler so that the eoc worker could be
started.

CRs-Fixed: 393301
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
(cherry picked from commit 7f1d97fd69c3afe66577e731105207e5db544ab2)

Change-Id: I1f754c34aeece255dafb1c1b2f07b9124d082bd5
Signed-off-by: Sridhar Gujje <sgujje@codeaurora.org>