msm: cpr: Allow setting previous voltage after frequency transition
As per the regulator framework, regulator_set_voltage is a no-op if
the new voltage is same as the previous. Use regulator_sync_voltage
instead, which forcefully applies the same voltage. This is required
as there is a possibility of the voltage being changed by an external
source without the knowledge of the regulator framework.

Signed-off-by: Kaushal Kumar <kaushalk@codeaurora.org>
(cherry picked from commit 5323ddc592fcd5c870d9c4e432aecd918d5c0c2b)

Change-Id: If805e00096b2b99487fc6bba9736abee037e9e80
Signed-off-by: Utsab Bose <ubose@codeaurora.org>