msm: acpuclock: Update jiffies while doing intermediate freq. switches
It is important to update the jiffies while doing the intermediate
frequency switches so that delay gets calculated correctly. If
not done then it is possible that some of the slow devices
would fail to transition to highest frequencies.

Signed-off-by: Trilok Soni <tsoni@codeaurora.org>
(cherry picked from commit 613d2d8720018071a6f1c62e73ffc27c924cb68f)

Change-Id: I0621a0271f8a26a976efb58e9a4528e7a3e7d688
Signed-off-by: Utsab Bose <ubose@codeaurora.org>
Signed-off-by: Sridhar Gujje <sgujje@codeaurora.org>