usb: gadget: Register RMNET SMD control channel only on protocol connect
The RMNET USB function driver uses SMD channel for transmitting control
messages to the modem processor. Move the registration of the platform
driver which handle the SMD channel to the stage in which the RMNET
function driver is actually got activated by the host which selected
the configuration which contain the RMNET function driver.

This fix is required to prevent cases in which other protocols, as
MBIM for example, which exists in other USB configuration, may need
the same SMD channel which is occupied by the RMNET, since the RMNET
grab the channel on initialization of the function driver rather than
upon activating it.

(cherry picked from commit b84439b31b9fc9ffdad46a4add8dd451aa2c7e0d)

Change-Id: Ie02bb25c56633e6cfafcc058fc17477ab4f2354c
CRs-Fixed: 435979
Signed-off-by: Ido Shayevitz <idos@codeaurora.org>