msm: mdss: move vsync pending logic for faster panel on
If vsync control from user space is called before panel is turned on,
this needs to be triggered as soon as panel is on. Since panel can also
be turned on in other places such as unblank, need to move this logic to
a common place as soon as panel is turned on faster.

CRs-Fixed: 452194
Change-Id: Ic8cf97df6af465b21610f5618f7eae76caa8423e
Signed-off-by: Adrian Salido-Moreno <adrianm@codeaurora.org>