msm: camera: Add support for IDEAL RAW.
When IDEAL RAW is used in ISP, the isp path should be
enabled instead of the bus path which streams the image to
the bus directly. This change fix the path configuration
when IDEAL RAW is in used. Also added configuration to support
MIPI raw, PLAIN8 and PLAIN16 RAW through IDEAL RAW interface.

Change-Id: I7f8b12ac45f4a9db981d5f7d57bf6ea5f4665f4a
Signed-off-by: Peter Liu <pingchie@codeaurora.org>
Signed-off-by: Sunil Joseph <jsunil@codeaurora.org>