usb: gadget: rmenet: Remove QTI block on usb cable disconnection
QTI (Qualcomm Tethering Interface) is a user space module that transfers
control packets between USB and modem.

On usb cable disconnection QTI was blocked, and didn't receive a
notification on the cable disconnection, so QTI couldn't notify the modem
on this event as well. Once we connect the USB cable the QTI receive both
disconnect message and connect message and notify the modem. Since the
modem couldn't handle those event consecutively, the connect event got
lost.

By this fix, we don't block the QTI on cable disconnection, so USB will
notify QTI on the disconnection leaving it and the modem enough time to
handle the disconnection event.

CRs-Fixed: 464298
Change-Id: I212b316e6aee1753f2f7f1bb9258653fa9d9bf41
Signed-off-by: Shimrit Malichi <smalichi@codeaurora.org>