usb: gadget: u_bam: fix race condition
On a usb cable connection/disconnection we place in a work queue the
function that is responsible to handle the connection/disconnection for
the RMNET interface.

If we connect the usb cable and disconnect it, we might get to the
condition that the connection function is running, though the usb cable
is disconnected. This may result in a NULL pointer access.

By this provided fix we check if the usb cable is disconnected before
trying to access a parameter that is set to NULL on cable disconnection.

CRs-Fixed: 468198
Change-Id: I5820c142ef84a7bfde773c725f6c3241f4ad60d6
Signed-off-by: Shimrit Malichi <smalichi@codeaurora.org>