usb: gadget: Support power collapse for cable disconnect
This change adds support for power collapse for tethering
functions of the USB device for cable connect/disconnect
scenarios.
The implication of this is that upon cable connect/disconnect
USB goes into low power mode, after handling a handshake via
resource manager with the IPA-A2 for graceful disconnection
of the pipes.

Change-Id: Ieb16812a4b9433d383862ff617d1f247c389f6ac
Signed-off-by: Lena Salman <esalman@codeaurora.org>