ASoC: mdm9625: Fix Voice over USB on MDM
USB driver uses AFE proxy ports in Audio for Voice over
USB feature. For this to support from Audio side the PCM
device number for AFE proxy ports have to be same across
targets. So the AFE proxy ports are arranged in correct
order in machine driver for USB driver to correctly use
it.
Change-Id: If3aee48ff7cad963e285a43e02116c259ed56270
Signed-off-by: Venkat Sudhir <vsudhir@codeaurora.org>