ARM: dts: msm: Use phandles for PM8941 local slave IDs
Add a pair of phandles to represent PM8941 local slave IDs 0 and
1.  The global slave ID used in SPMI transactions corresponding
to these local slave IDs changes depending upon whether or not
PM8941 is the primary PMIC in the system.

Using phandles for the local slave IDs allows the global slave
IDs for the local slave IDs to be specified on a per board basis.
This is required for MSM8974Pro-AC where the global slave IDs for
PM8941 change.

Change-Id: If2af21b8069d73a602da61cc000da602164b3442
Signed-off-by: David Collins <collinsd@codeaurora.org>
Signed-off-by: Shashank Babu Chinta Venkata <sbchin@codeaurora.org>