ARM: dts: msm: Add various SPMI peripheral devices for PMIC pma8084
Add device nodes for the following SPMI peripherals to the PMIC
PMA8084 device tree file:
	1. 0x0100 - revid
	2. 0x0800 - power-on
	3. 0x0900 - misc
	4. 0x2400 - temp-alarm
	5. 0x2800 - coincell
	6. 0x6000 - rtc

PMA8084 and PM8941 share the six peripherals listed above.
Adding them for PMA8084 brings PMA8084 almost to feature parity
with PM8941.  This is important for MSM8974Pro-AC where PMA8084
is used as the primary PMIC in place of PM8941.

Change-Id: Iea6f9aefb37ca02b1164574458d6efec07f1c918
Signed-off-by: David Collins <collinsd@codeaurora.org>
Signed-off-by: Shashank Babu Chinta Venkata <sbchin@codeaurora.org>