usb: usb_bam: Don't wait for consumer request on connect pipes
On certain compositions, ECM for instance, consumer is not requested
on pipe connection and thus the part of the handshake that waits
for consumer request causes a timeout and a delay of 10 seconds
before network connection can be established.
In a revisited design for IPA-USB resource handshake,
USB shall not wait for USB_CONS resource, thus avoiding the
delay on ECM compositions.

Change-Id: I7fb281b871a7bc471d73a7d163c8f7b8861249a8
CRs-fixed: 499480
Signed-off-by: Devendra Patel <cdevenp@codeaurora.org>