arm/dt: msm8226: Add helper regulator for vdd_buck of audio codec
vdd_buck power rail to the codec is supplied by SMPA 4. During
high loading conditions, SMPA4 needs to be backed up by LDO25
as a helper regulator. Add LDO25 as a helper regulator to the
audio codec regulator list. Both SMPA4 and LDO25 are optional
and on-demand regulators for the audio codec.

Change-Id: I2b28d645d21f12c7c9e0badb9101ed7dab075dee
Signed-off-by: Bhalchandra Gajare <gajare@codeaurora.org>