ASoC: msm8x10-wcd: Initialize IRQ driver from codec
In order to request for interrupts, the codec driver needs
to initialize the codec IRQ driver. Add support to initialize
the IRQ driver with a codec specific interrupt table specifying
the interrupts supported by the codec and the order in which
to process the interrupts.

Change-Id: I9d9a286d07b411e04fd05cc9da9a31c755353b21
Signed-off-by: Bhalchandra Gajare <gajare@codeaurora.org>
(cherry picked from commit d7d779f7fd59352ca1fddb99b66faa7abf2287d0)

Signed-off-by: Gopal G Goberu <ggober@codeaurora.org>