ASoC: msm8x10-wcd: Initialize IRQ driver from codec.
In order to request for interrupts, the codec driver needs
to initialize the codec IRQ driver. Add support to initialize
the IRQ driver with a codec specific interrupt table specifying
the interrupts supported by the codec and the order in which
to process the interrupts.

Change-Id: I313a92fd43537c126e5582481072892b4a14aeb5
Signed-off-by: Bhalchandra Gajare <gajare@codeaurora.org>
Signed-off-by: Sanrio Alvares <sanrioalvares@codeaurora.org>
(cherry picked from commit d130c781c865e19ebd006283c1ef5fba85d8b70e)

Signed-off-by: Gopal G Goberu <ggober@codeaurora.org>