usb: msm_otg: Workaound for USB HW issue with remote-wakeup during LPM
When a remote-wakeup operation is issued by the SW during USB
low-power mode, the USB HW issues a J to K state transition on the USB bus.
Shortly after this transition the D- lines drops from 3.3V to 1.5V.
The SW workaround implemented for this issue is to write the FPR bit of
the PORTSC register (which triggers USB remote-wakeup) in the same write
of PHCD bit clearing (which exits the USB HW from low-power mode).

Change-Id: I0cd12ca131779d6a610956d5be22fe3be194964f
CRs-Fixed: 502579
Signed-off-by: Daniel Segal <dsegal@codeaurora.org>