msm: kgsl: Hold a reference to process private when creating context
When a context is created hold a reference to the process private
structure which is referenced from within the context. The process
private structure can be destroyed while some threads still have
a reference to the context, this can lead to a situation where the
process private pointer inside the context pointer becomes invalid.
Avoid this by holding a reference to the process private structure
as long as the context is around.

Change-Id: Ia35629e5d027a383ed4c1378316633b4923372f7
Signed-off-by: Shubhraprakash Das <sadas@codeaurora.org>