USB: ipc_bridge: Fix a hang in the disconnect path
The reader process is woken up when a read urb is completed.  The
read urb may not be active during disconnect.  So killing it will
not wake up the reader process.  The client driver is blocked in
flushing the read work due to this reason.  Fix it by explicitly
waking up the reader process in disconnect path.

Change-Id: Ibb698eae3d81d936f303a58335065367e724324f
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>