msm: kgsl: Check whether context is detached before updating timestamp
When context is detached there is a wait for the last timestamp
used by the context. This is done under the assumption that after
this timestamp there will be no commands in the pipeline that are
submitted on behalf of this context. However, the dispatcher can
still dispatch commands from the detached context because there
is a time gap between the context being detached and the dispatcher
checking whether the context is detached. Hence, guard the last
timestamp by holding the context mutex and checking if the context is
detached or not before updating it.

Change-Id: I780372a1b2c2ddb4772310918ba28118740d85ef
Signed-off-by: Shubhraprakash Das <sadas@codeaurora.org>