usb: msm_otg: wake from apps power collapse during bus resume
In case D+ and D- MPM pins exists in the platform, turn on/off
the level shifter that drive these signals into the MPM as well
as the corresponding MPM pins.

This will allow the chip to wake up from apps power collapse by
the MPM when the USB cable is connected and USB bus resume occurred.

Change-Id: I4785f73dd45e69eb389fa812f7aec8fa48c756c5
Signed-off-by: Ido Shayevitz <idos@codeaurora.org>
Signed-off-by: Amit Blay <ablay@codeaurora.org>