ARM: dts: msm: enable D+ HV interrupt for 9x25
On 9x25, the USB D+ and D- signals are routed from the HSUSB PHY to
dedicated MPM pins. This way the MPM can wake the chip  when these
signals change.

In this commit we enable only the D+ MPM pin, since we would like the MPM
to wake the chip on USB bus resume, which happens when the D+ is dropping.

Change-Id: I2b1075cc97fa9a1004aafedfc59a9eb9ecc70035
Signed-off-by: Ido Shayevitz <idos@codeaurora.org>
Signed-off-by: Amit Blay <ablay@codeaurora.org>