msm: msm_bus: Move counting of bus devices to device-tree
So far the bus driver has been getting the number of bus devices
using a macro defined in the board file.
Changing this to leverage device-tree.
Now the number of bus devices can be read from the device tree
reducing the number of code changes required for adding new
targets.

Change-Id: I56c1a056eb9f7a85c9fb295b2b25e4d0f7896eef
Signed-off-by: Gagan Mac <gmac@codeaurora.org>