mdss: mdp3: Increase memory mapping of buffers used by mdp3
There is a hw issue with ppp where if the roi is not divisible
by 16 it is possible that hw will try to fetch memory that is
outside image region. This additional read is not used in
composition. We are simply mapping a dummy buffer at the end
to avoid ppp to read the data which is not mapped.

Change-Id: Ib471e757ba6f5028affdbbaea7a24f5245b1103d
Signed-off-by: Terence Hampson <thampson@codeaurora.org>