msm: mdss: Correct the data type and boundary checking for HSIC
For HSIC matrix vectors are define as unsigned integers. But they
can get negative values as well. In this case the limit checking
for this fails and ioctl returns error failing to set HSIC values.
This patch corrects the data type for matrix vectors and fine
tunes the boundary values.

CRs-fixed: 535931
Change-Id: Ia84ac6574d8b7004d2ead499e722e5da1ccd4411
Signed-off-by: Shivaraj Shetty <shivaraj@codeaurora.org>