Revert "debugfs: more tightly restrict default mount mode"
This reverts commit 82aceae4f0d42f03d9ad7d1e90389e731153898f.

The wlan module makes some debugfs directories when it is loaded.
Android userspace loads the wlan module as the system user and
the system user doesn't have write access to the debugfs
filesystem because the default permissions for debugfs are 0700
now. Revert this patch to allow the wlan module to create
directories in debugfs and successfully probe.

Change-Id: I79c9598dcdf83bc5d418c281b0eedea5eae298f6
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>