usb: msm_otg: Fix BAM remote-wake up issue during USB low-power mode
When new data is pushed into the USB BAM pipes during USB low-power mode,
the BAM wakeup ISR triggers a USB remote-wakeup operation. This leads to
calling the ci13xxx_wakeup() function. This function issued a low-power
mode exit by a remote task, and then sets the FPR bit for issuing remote
wakeup. The problem is that sometimes the low-power mode exit remote task
does not get a chance to be executed before setting the FPR bit. In result,
the write to FPR bit fails, and no remote-wakeup happens.
The solution implemented for this issue is to wait for the low-power
mode exit to complete before setting the FPR bit.

Change-Id: I1b2ad94cc5943dd2365dba7c4567d1930662dca3
CRs-Fixed: 514368
Signed-off-by: Danny Segal <dsegal@codeaurora.org>
Signed-off-by: Tarun Gupta <tarung@codeaurora.org>