msm: soc: add machine name check for msm8x28
Here change the machine_name when the soc is msm8x28 in setup_arch().
But it will involve the msm related API into the arm boot code in
arch/arm/kernel/setup.c.

Another sulution:
1. Add another DT_MACHINE_START for 8228
2. Select the 8228 machine in setup_machine_fdt() - devtree.c
3. But it will broke the normal machine parsing/matching process. In
another words, we need add msm specific check in common DT parsing process,
which is not accepted.

So adopt the first solution here as a trade-off.

Change-Id: I8693ddda888340ccbc19f4f528fe766ace9e9689
Signed-off-by: Xiaocheng Li <lix@codeaurora.org>