ASoC: wcd9xxx: Enable MBHC TxFE when micbias is enabled
When micbias is turned ON for audio recording with headset
plugged in, there can be possible noise on the recorded
data due to micbias polling for button detection. To
avoid the noise on audio recording, it is required to
enable the MBHC TxFE (Front-End). Add support to enable
MBHC TxFE when micbias is being enabled for audio recording.

Change-Id: Ic9dee6dd139848dd658d5c382a9ef25b4d801305
CRs-fixed: 522829
Signed-off-by: Bhalchandra Gajare <gajare@codeaurora.org>