msm: qmi: Update the skip_to_next_elem operation
skip_to_next_elem operates based on tlv_type and all elements in
nested structures have the same tlv_type. If the length of a variable
length array element in a nested structure is zero, then this operation
skips the entire nested structure instead of moving to the next element
in the nested structure.

Fix the skip_to_next_elem operation for nested structure by moving to
next element.

Change-Id: I95fbf436598f969aae5b850264b81ff817cb5d28
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@codeaurora.org>