usb: dwc3-msm: Use HSPHY and SSPHY drivers
Routines to initialize and suspend the USB PHYs are now handled
in their own drivers, which simplifies the controller/glue from
having to do it centrally. Change the DWC3 MSM glue layer to use
USB PHY API to make such calls. The PHY instances will be attached to
the controller in the device tree, which are referenced by their
phandles.

Change-Id: I05c3664d09e0b0f3fceb2b62aeb8d06a2cc18843
Signed-off-by: Jack Pham <jackp@codeaurora.org>