msm: mdss: Implement overlay handoff for continuous splash screen
In order to support a smooth and seemless transition from the splash
screen to the boot animation, MDP needs to handoff all the
hardware ctl, mixers and pipes that were staged by the bootloader.
Implement this functionality by populating the MDP software structure
to reflect the state of the hardware at probe time. This removes the
the need to turn off the MDP timing generator when transitioning to
the boot animation and reduces this delay to just one frame.

Change-Id: I48f733cfaa3112eb6995e2fc70577f9defaf4600
Signed-off-by: Aravind Venkateswaran <aravindh@codeaurora.org>