usb: gadget: ci13xxx: ensure exit from lpm on software connect/disconnect
In case the USB gadget controller is in low power mode (lpm) due
to bus suspend, the controller may not be accessible, however
the gadget framework still allow software disconnect and connect using the
pullup operation. This may results with a crash when we try to access
the controller.

Therefore we must resume the controller and exit lpm before do any
write/read operations to the controller registers.
Write helper function to accomplish that and use it upon pullup operation.

CRs-Fixed: 550217
Change-Id: I2c827b241880b1ea9db1c0bd870b16ea9f0b9ca5
Signed-off-by: Ido Shayevitz <idos@codeaurora.org>