msm: vidc: Handle SYS_IDLE interrupt concurrency
It is possible to receive SYS_IDLE message before
Venus core sends the interrupt for that message.
This results in clocks not being disabled at all
and device will continue to drain power.

This change checks when the interrupt is received
if the venus core is still in IDLE state. If it is
then disable the clocks.

Change-Id: I4122139df67179ccdfe0d1b175c17c69ddc995cc
CRs-fixed: 548745
Signed-off-by: Sachin Shah <sachins@codeaurora.org>
Signed-off-by: Abhishek Ranjan <aranja@codeaurora.org>