diag: Forward pkts to peripherals only after receiving feature mask
Diag feature masks are exhanged between the Apps processor and the
peripherals during boot up that contains the information about the
channel used for transmitting/receiving command responses. Do not
process any command until we receive the feature masks from the
peripherals. Also continue to process the ctrl pkt even if the
registration table count is 0 to process the feature mask at the end
of the ctrl pkt.

Change-Id: Iabd6e7dfe803e3f6ae08b287d68a023cef40f6cc
CRs-Fixed: 550023
Signed-off-by: Ravi Aravamudhan <aravamud@codeaurora.org>