usb: dwc3_otg: Remove phy dependency from xhci-plat
xhci-plat is generic code that should not be tied to either OTG
or DWC3-specific implementations. The hook to otg_set_host()
can be handled by DWC3's OTG driver itself when adding/removing
the xHCI platform device.

Furthermore, the runtime PM callbacks that call PHY's set_suspend()
can also be eliminated, as the PHYs are being suspended by the
DWC3 core and glue layers. This will happen automatically so long
as the parent-child relationship is maintained. As a result, remove
dwc3_otg_set_suspend as it is no longer used.

Change-Id: I62e98693fbb67739ef4d2ea40b12feb0d3795ff4
Signed-off-by: Jack Pham <jackp@codeaurora.org>