misc: Add snapshot of Liquid Docking Station driver
Add a snapshot of the Liquid Docking Station driver
(formerly known as the APQ8084 Liquid Dock driver) as of
commit 156aa1e4fe6a6f3a71275ebabd9a75024494405c
('Merge "ARM: dts: msmzirc: Enable USB device"').

This driver is used to coordinate the GPIOs used to power
up and reset the USB hub and ethernet ports on the QTI
Liquid Docking Stations.

Change-Id: Ie1baf294396086aaaa342025349aa146bef8d0b5
Signed-off-by: Jack Pham <jackp@codeaurora.org>
[stepanm@codeaurora.org: resolve Makefile conflicts]
Signed-off-by: Stepan Moskovchenko <stepanm@codeaurora.org>