net: usb: Add IPA datapath support to USB Ethernet adapters
In order to support high throughput requirements, USB ethernet
adapter datapath can be sent via IPA hardware. This change adds
support for USB ethernet adapter to open a BAM2BAM pipes with
the IPA, and cancel the need for the packets to go through SW.
The control over exception packets shifts to the IPA-ODU
network device, and the usual usbnet device is not used.

Change-Id: I9d3f9cc7efd01508ced9c9a8be6024c8038d7e4d
Signed-off-by: Lena Salman <esalman@codeaurora.org>