msm: memshare: Handle modem subsystem restart event
Currently once the memory is allocated to modem, re-allocation
of memory cannot be done unless it is freed up. After memory
allocation if modem subsytem restart happens Modem is not
aware of any memory being allocated to it earlier. Thus if
modem again request for memory allocation, without freeing up
earlier allocated memory, we will pass the same handle as of
earlier.

Change-Id: I88869088cad6c3db8713dd4c1589821b42ecff46
Signed-off-by: Katish Paran <kparan@codeaurora.org>