USB: gadget: Fix IPA disconnect on suspend change
The IPA disconnect on suspend change has introduced a new is_connected flag
for tracking the BAM port state. However, this flag is not updated on the
connect and disconnect work functions. With this fix it is set to true in
the end of the connect work function and to false in the end of the
disconnect work functions.

Change-Id: Ibe22929b06038bee8c3b29e01cc5b25008758935
CRs-fixed: 571703
Signed-off-by: Danny Segal <dsegal@codeaurora.org>