ARM: dts: msm: Change the capless mode for 8226 qrd
msm8226-qrd.dtsi is included by many 8226 or 8926 dts,
but not all the boards have the external cap for micbias1.
For example, 8926 qrd board doesn't have the cap.
So remove the external cap for micbias1 in this dtsi
and add it into the specific board dts who uses the cap.

CRs-fixed: 572412

Change-Id: I196657fabbab46ff6ff8d36b25554da50fab9f16

Signed-off-by: Kunlei Zhang <quic_kunleiz@codeaurora.org>