msm: smd: Add probe deferral support
SMD APIs may be called before probe() has a had a chance to initialize
the driver.  In such cases, EPROBE_DEFER should be returned to the caller
to indicate SMD needs more time to probe.

Change the public APIs to detect the current status of probe() and return
EPROBE_DEFER if necessary.

Change-Id: I243f56c3aa6afdce22192f17bcbae9377242d116
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
[joshc: drop all but msm_smd_pkt.c change]
Signed-off-by: Josh Cartwright <joshc@codeaurora.org>